Added new properties in the Car class: a fuel type that the car uses, a timer that the car will fuel for which is used to check it has been fuelled, a boolean to check if a car has fuelled and the pump the car is at. The timer's properties are adjusted in the Main method to only elapse once.
Another array has been created to store cars that have left and a jagged array is used to store all the pumps in their respective lanes instead of individual variables.

a new timer is added called Assigner, this checks every 100ms to see if a car that is not fuelling can be fuelled, it is then allocated a pump that is farthest away for efficiency reasons. Its properties are checked to allow it to be seen in a fuelling state before it is sent to another array.

Comments were added on other existing loops to demonstrate their capabilities.

The fuelling time timer would check if a car has finished fuelling and reset the pump so it can be used again, it will also store the running total of fuel that it has dispensed for each fuel type.

In the petrol class, the fuel dispensed are now doubles and I added an unleaded dispensed property.